# Conflicts:
#   frontend/.env.example

+++ backend/src/main/java/com/conkiri/domain/base/repository/SeatRepository.java
@@ -9,6 +9,8 @@ import com.conkiri.domain.base.entity.Seat;
 import com.conkiri.domain.base.entity.Section;
 
 public interface SeatRepository extends JpaRepository<Seat, Long> {
+
 	List<Seat> findBySection(Section section);
-	Optional<Seat> findByRowLineAndColumnLineAndSection(Integer rowLine, Integer columnLine, Section section);
+
+	Optional<Seat> findByRowLineAndColumnLineAndSection(Long rowLine, Long columnLine, Section section);
 }
